:truck: [Rename]: ResponseDTO response 디렉토리로 이동
